+++ app.js
@@ -1,5 +1,6 @@
 import express from "express";
 import db from "./database.js"; // Needed to start database on run
+import cors from "cors";
 
 const app = express();
 const port = 3000;
@@ -7,6 +8,8 @@ const port = 3000;
 // Route imports
 import user from './routes/user.js';
 import server from './routes/server.js';
+import conversation from './routes/conversation.js';
+import request from './routes/request.js';
 
 // Start backend app
 app.listen(port, () => {
@@ -14,9 +17,16 @@ app.listen(port, () => {
 });
 
 // middleware
+app.use(cors({
+  'allowedHeaders': ['Content-Type'],
+  'origin': '*',
+  'preflightContinue': true
+}));
 app.use(express.json());
 app.use(express.urlencoded());
 
 // Routes
 app.use('/user', user)
 app.use('/server', server);
+app.use('/conv', conversation);
+app.use('/req', request);