+++ models/user.schema.js
@@ -2,14 +2,19 @@ import mongoose from "mongoose";
 import {seq} from "../database.js";
 
 const userSchema = new mongoose.Schema({
+    uid: {type: Number, index: true, unique: true},
     username: String,
-    password: String
+    password: String,
+    friends: {type: [Number], default: []}
 });
 
-userSchema.plugin(seq, {inc_field: "uid", start_seq: 4, inc_amount: 3});
+userSchema.plugin(seq, {inc_field: "uid", start_seq: 100, inc_amount: 3});
 
 userSchema.pre("save", function(next) {
     const doc = this;
+    console.log(doc);
+
+    if (doc._id) next();
 
     User.find({username: doc.username}, function(err, docs) {
         if (!docs.length) next();
@@ -20,7 +25,7 @@ userSchema.pre("save", function(next) {
             next(e);
         }
     });
-})
+});
 
 const User = mongoose.model("User", userSchema);
 export default User;